pyproject.toml
[tool.poetry]
name = "youtube-speak"
version = "0.1.0"
description = ""
authors = ["egeres <egeres.default@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
rich = "^13.3.5"
jstyleson = "^0.0.2"
json5 = "^0.9.14"
pytube = "^15.0.0"
jsonargparse = "^4.27.0"
webvtt-py = "^0.4.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
pre-commit = "^3.3.1"
tomlkit = "^0.11.8"
pydocstyle = "^6.3.0"
json5 = "^0.9.11"
coverage = "^7.2.5"
hypothesis = "^6.75.3"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
